+++ xs/sources/syntax/parser_declaration.c
@@ -237,15 +237,6 @@ static XsSyntaxNode *parse_function(SyntaxParser *parser, Modifiers modifiers, s
     return_type->flags |= XS_SYNTAX_FLAG_RETURN_TYPE;
     xs_syntax_node_add(parser->tree, function, return_type);
   }
-  if(accept(parser, XS_TOKEN_KW_THROWS))
-  {
-    xs_diagnostics_add(parser->diagnostics, XS_DIAGNOSTIC_WARNING, parser->previous.span,
-                       "exception syntax is deprecated; prefer Result<T, E>");
-    do
-    {
-      xs_syntax_node_add(parser->tree, function, parse_type(parser));
-    } while(accept(parser, XS_TOKEN_COMMA));
-  }
   if(accept(parser, XS_TOKEN_SEMICOLON))
   {
     if(!signature_allowed && (function->flags & XS_SYNTAX_FLAG_INCOMPLETE) == 0)